# packages/protocol/openapi/agentnet.openapi.yaml
openapi: 3.1.0
info:
  title: AgentNet API
  description: AgentNet is a centralized Agent Relay Platform for registering agents,
    routing asynchronous tasks, tracking delivery state, and enforcing cross-agent
    approval policy. WebSocket agent transport is available at `/v1/ws` and documented
    in `docs/openapi.md`.
  version: 0.1.0
paths:
...
  /v1/dashboard/admin/dedicated-channels:
    get:
      tags:
      - dashboard
      summary: List Dedicated Channels
      description: 'List dedicated channels with optional filtering (admin read).


        Requires: admin:read

        Secrets in connection_config and encryption_config are masked.

        Audit: records every read of sensitive channel configuration metadata.'
      operationId: list_dedicated_channels_v1_dashboard_admin_dedicated_channels_get
      parameters:
      - name: channel_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by channel type
          title: Channel Type
        description: Filter by channel type
      - name: source_agent_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Filter by source agent
          title: Source Agent Id
        description: Filter by source agent
      - name: target_agent_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Filter by target agent
          title: Target Agent Id
        description: Filter by target agent
      - name: enabled
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          description: Filter by enabled status
          title: Enabled
        description: Filter by enabled status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: agentnet_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agentnet Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DedicatedChannelListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyBearer: []
      - LegacyApiKeyHeader: []
    post:
      tags:
      - dashboard
      summary: Create Dedicated Channel
      description: 'Create a new dedicated channel (super_admin + step-up required).


        Requires: super_admin:write + step-up'
      operationId: create_dedicated_channel_v1_dashboard_admin_dedicated_channels_post
      parameters:
      - name: agentnet_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agentnet Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DedicatedChannelCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DedicatedChannelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyBearer: []
      - LegacyApiKeyHeader: []
  /v1/dashboard/admin/dedicated-channels/{channel_id}:
    get:
      tags:
      - dashboard
      summary: Get Dedicated Channel
      description: 'Get a single dedicated channel by ID (admin read).


        Requires: admin:read'
      operationId: get_dedicated_channel_v1_dashboard_admin_dedicated_channels__channel_id__get
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      - name: agentnet_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agentnet Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DedicatedChannelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyBearer: []
      - LegacyApiKeyHeader: []
    patch:
      tags:
      - dashboard
      summary: Update Dedicated Channel
      description: 'Update a dedicated channel (super_admin + step-up required).


        Requires: super_admin:write + step-up'
      operationId: update_dedicated_channel_v1_dashboard_admin_dedicated_channels__channel_id__patch
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      - name: agentnet_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agentnet Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DedicatedChannelUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DedicatedChannelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyBearer: []
      - LegacyApiKeyHeader: []
    delete:
      tags:
      - dashboard
      summary: Delete Dedicated Channel
      description: 'Delete (revoke) a dedicated channel (super_admin + step-up required).


        Requires: super_admin:write + step-up'
      operationId: delete_dedicated_channel_v1_dashboard_admin_dedicated_channels__channel_id__delete
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      - name: agentnet_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agentnet Session
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyBearer: []
      - LegacyApiKeyHeader: []
  /v1/dashboard/admin/dedicated-channels/{channel_id}/enable:
    post:
      tags:
      - dashboard
      summary: Enable Dedicated Channel
      description: 'Enable a dedicated channel (super_admin + step-up required).


        Requires: super_admin:write + step-up'
      operationId: enable_dedicated_channel_v1_dashboard_admin_dedicated_channels__channel_id__enable_post
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      - name: agentnet_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agentnet Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DedicatedChannelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyBearer: []
      - LegacyApiKeyHeader: []
  /v1/dashboard/admin/dedicated-channels/{channel_id}/disable:
    post:
      tags:
      - dashboard
      summary: Disable Dedicated Channel
      description: 'Disable a dedicated channel (super_admin + step-up required).


        Requires: super_admin:write + step-up'
      operationId: disable_dedicated_channel_v1_dashboard_admin_dedicated_channels__channel_id__disable_post
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      - name: agentnet_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agentnet Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DedicatedChannelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyBearer: []
      - LegacyApiKeyHeader: []
...
components:
  schemas:
...
    DedicatedChannelCreate:
      properties:
        scope_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Scope Id
          description: Network scope ID
        channel_name:
          type: string
          maxLength: 128
          minLength: 1
          title: Channel Name
          description: Human-readable channel name
        channel_type:
          type: string
          title: Channel Type
          description: Channel type (vpn, private_link, p2p, direct_connect)
        source_agent_id:
          type: string
          format: uuid
          title: Source Agent Id
          description: Source agent ID
        target_agent_id:
          type: string
          format: uuid
          title: Target Agent Id
          description: Target agent ID
        connection_config:
          type: object
          title: Connection Config
          description: Connection configuration
        encryption_config:
          type: object
          title: Encryption Config
          description: Encryption configuration
        bandwidth_mbps:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Bandwidth Mbps
          description: Target bandwidth in Mbps
        latency_target_ms:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Latency Target Ms
          description: Target latency in ms
      type: object
      required:
      - channel_name
      - channel_type
      - source_agent_id
      - target_agent_id
      title: DedicatedChannelCreate
      description: Request body for creating a dedicated channel.
    DedicatedChannelListResponse:
      properties:
        channels:
          items:
            $ref: '#/components/schemas/DedicatedChannelResponse'
          type: array
          title: Channels
        total:
          type: integer
          title: Total
        offset:
          type: integer
          title: Offset
        limit:
          type: integer
          title: Limit
      type: object
      required:
      - channels
      - total
      - offset
      - limit
      title: DedicatedChannelListResponse
      description: Paginated list of dedicated channels.
    DedicatedChannelResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        scope_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Scope Id
        channel_name:
          type: string
          title: Channel Name
        channel_type:
          type: string
          title: Channel Type
        source_agent_id:
          type: string
          format: uuid
          title: Source Agent Id
        target_agent_id:
          type: string
          format: uuid
          title: Target Agent Id
        connection_config:
          type: object
          title: Connection Config
        encryption_config:
          type: object
          title: Encryption Config
        bandwidth_mbps:
          anyOf:
          - type: number
          - type: 'null'
          title: Bandwidth Mbps
        latency_target_ms:
          anyOf:
          - type: number
          - type: 'null'
          title: Latency Target Ms
        enabled:
          type: boolean
          title: Enabled
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - scope_id
      - channel_name
      - channel_type
      - source_agent_id
      - target_agent_id
      - connection_config
      - encryption_config
      - bandwidth_mbps
      - latency_target_ms
      - enabled
      - created_at
      - updated_at
      title: DedicatedChannelResponse
      description: Single dedicated channel response.
    DedicatedChannelUpdate:
      properties:
        channel_name:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 1
          - type: 'null'
          title: Channel Name
        connection_config:
          anyOf:
          - type: object
          - type: 'null'
          title: Connection Config
        encryption_config:
          anyOf:
          - type: object
          - type: 'null'
          title: Encryption Config
        bandwidth_mbps:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Bandwidth Mbps
        latency_target_ms:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Latency Target Ms
      type: object
      title: DedicatedChannelUpdate
      description: Request body for updating a dedicated channel.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    ApiKeyBearer:
      type: http
      scheme: bearer
      description: 'REST API key in `Authorization: Bearer ak_...`.'
    LegacyApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: Legacy REST API key header.
...
servers:
- url: http://localhost:8000
  description: Local development